fix valueOf implementation

+++ src/main/java/uoc/ded/practica/model/Group.java
@@ -65,12 +65,12 @@ public class Group {
     }
 
     public Double valueOf() {
-        int badgeValueSum = 0;
+        Double badgeValueSum = 0.;
 
         for (Iterador<User> it = this.members(); it.haySiguiente();) {
             badgeValueSum += it.siguiente().getBadge(this.date).getValue();
         }
-        Double averageBadgeValue = (double) (badgeValueSum / this.numMembers());
+        Double averageBadgeValue = badgeValueSum / this.numMembers();
         return averageBadgeValue;
     }
 }
